Migrating work done by Jake to main site
A secondary webserver was created holding webpages. These pages were built using an older version of Django. Currently updating the files to run on our version of Django.

***NOTE***
There is still an issue with the views.py file inside of the music app. Appears to be a connection issue to the models.py

Co-Authored-By: imgjoh <imgjoh@users.noreply.github.com>